feat: Implimented the `axios` library and introduced `promises`
The `axios get` method returns a `promise`, so we can access it directly using then. In the code I
use `then` once to do something with the geolactions data. I print the address ot the screen. Then I
return another promise, when I make the request for weather. Inside of the second `then` call, I
print the weather to the screen. I also addded a catch call, which will handle any errors. If
anything goes wrong with either of our promises or if we throw an error, catch will get fired
printing the messages to the screen.